Sumando los cambios del diego
ultimo sprint y sus cambios locos

+++ taa-2.0/taa-2.0-war/src/java/cl/usach/pingeso/taa2/managedbeans/ProgramDetailsMB.java
@@ -50,7 +50,7 @@ public class ProgramDetailsMB {
         }
         this.code = programSelect.getProgramCode();
         this.name = programSelect.getProgramName();
-        this.levels = programSelect.getLevels();
+        this.levels = programSelect.getLevels().shortValue();
     }
     
     public void backToPrograms() {
